+++ CHANGELOG.md
@@ -52,6 +52,7 @@
 * Adjusted Flex table component to make it responsive to content. Refs [UILD-652].
 * Fixed shifted Admin metadata tooltip. Refs [UILD-654].
 * Set or delete the default user profile when changing the profile for a Work/Instance. Refs [UILD-645].
+* Change work Profile after it is created. Refs [UILD-658].
 
 [UILD-552]:https://folio-org.atlassian.net/browse/UILD-552
 [UILD-544]:https://folio-org.atlassian.net/browse/UILD-544
@@ -105,6 +106,7 @@
 [UILD-652]:https://folio-org.atlassian.net/browse/UILD-652
 [UILD-654]:https://folio-org.atlassian.net/browse/UILD-654
 [UILD-645]:https://folio-org.atlassian.net/browse/UILD-645
+[UILD-658]:https://folio-org.atlassian.net/browse/UILD-658
 
 ## 1.0.5 (2025-04-30)
 * Fixed incorrect behavior when navigating between duplicated resources. Fixes [UILD-553].